+++ website/src/lib/server-actions/campaigns-actions.ts
@@ -1,9 +1,23 @@
 'use server';
 
 import { getSessionByType } from '@/lib/firebase/current-account';
+import type { WebsiteLanguage } from '@/lib/i18n/utils';
 import { CampaignFormCreateInput, CampaignFormUpdateInput } from '@/lib/services/campaign/campaign-form-input';
 import { services } from '@/lib/services/services';
 import { revalidatePath } from 'next/cache';
+
+export const getCampaignPageContentAction = async (lang: WebsiteLanguage) => {
+	return services.read.campaignPublicWebsite.getPageContent(lang);
+};
+
+export const getCampaignPageMetadataAction = async (slug: string, lang: WebsiteLanguage) => {
+	const result = await services.read.campaign.getBySlug(slug);
+	if (!result.success || !result.data?.isActive) {
+		return services.read.campaignPublicWebsite.getFallbackMetadata(lang);
+	}
+
+	return services.read.campaignPublicWebsite.getPageMetadata(lang, result.data);
+};
 
 export const createCampaignsAction = async (campaigns: CampaignFormCreateInput) => {
 	const sessionResult = await getSessionByType('user');
